backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["SmokierLemur51 <ldl6147@gmail.com>"]
readme = "README.md"
packages = [{include = "backend", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.3"
flask-sqlalchemy = "^3.1.1"
flask-bcrypt = "^1.0.1"
flask-wtf = "^1.2.1"
python-dotenv = "^1.0.1"
flask-login = "^0.6.3"

[tool.poetry.scripts]
start = "backend:run"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
